raspberry-pi/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
paho-mqtt==1.6.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
eventlet==0.33.3
